test(sync): split m4 timer/webhook tests into two files
Split the pre-existing post-merge-validator-m4.test.ts (720 LOC) along
its natural axis into two concern-named files, each <=500 LOC:

- post-merge-validator-m4-timer-config.test.ts (438) — normalizeTimerConfig
  edge cases + runAllValidations integration paths exercising the timer
  validator (including timer-error-does-not-block-webhook).
- post-merge-validator-m4-webhook-config.test.ts (389) — normalizeWebhookConfigs
  validation (URL/protocol/event-types) + runAllValidations integration
  paths exercising the webhook validator (including both-validators and
  webhook-error-does-not-block-timer).

Cross-validator runAllValidations tests carry a brief comment noting the
cross-coverage at their landing site.
