+++ packages/sync/src/__tests__/post-merge-validator-m4-timer-config.test.ts
@@ -1,8 +1,12 @@
 /**
- * M4 post-merge validator tests: timer config and webhook config normalization.
+ * M4 post-merge validator tests: timer config normalization.
  *
- * Covers normalizeTimerConfig edge cases, normalizeWebhookConfigs validation,
- * and runAllValidations integration for both M4 document fields.
+ * Covers normalizeTimerConfig edge cases (interval bounds, waking-hours
+ * validity, archived skip) and runAllValidations integration paths that
+ * exercise the timer validator.
+ *
+ * Companion file: post-merge-validator-m4-webhook-config.test.ts holds
+ * normalizeWebhookConfigs tests + runAllValidations webhook integration.
  */
 import * as Automerge from "@automerge/automerge";
 import { WasmSodiumAdapter } from "@pluralscape/crypto/wasm";
@@ -13,17 +17,12 @@ import {
   createSystemCoreDocument,
   fromDoc,
 } from "../factories/document-factory.js";
-import {
-  normalizeTimerConfig,
-  normalizeWebhookConfigs,
-  runAllValidations,
-} from "../post-merge-validator.js";
+import { normalizeTimerConfig, runAllValidations } from "../post-merge-validator.js";
 import { EncryptedSyncSession } from "../sync-session.js";
 
 import { asSyncDocId, asTimerId } from "./test-crypto-helpers.js";
 
-import type { CrdtTimer } from "../schemas/system-core.js";
-import type { SystemCoreDocument } from "../schemas/system-core.js";
+import type { CrdtTimer, SystemCoreDocument } from "../schemas/system-core.js";
 import type { DocumentKeys } from "../types.js";
 import type { SodiumAdapter } from "@pluralscape/crypto";
 
@@ -74,36 +73,32 @@ function createTimerSession(
   });
 }
 
-/**
- * Helper to create a session with the "webhookConfigs" field for testing.
- * The SystemCoreDocument type does not include webhookConfigs yet, so we
- * build a minimal document via fromDoc with the correct shape.
- */
-function createSessionWithWebhookConfigs(
-  keys: DocumentKeys,
-  docId: string,
-): EncryptedSyncSession<WebhookTestDocument> {
-  const base = fromDoc({
-    timers: {},
-    webhookConfigs: {},
-  });
-  return new EncryptedSyncSession<WebhookTestDocument>({
-    doc: Automerge.clone(base),
-    keys,
-    documentId: asSyncDocId(docId),
-    sodium,
-  });
-}
-
 interface WebhookConfigShape {
   url: Automerge.ImmutableString;
   eventTypes: Automerge.ImmutableString[];
   enabled: boolean;
 }
-
-interface WebhookTestDocument {
+interface TimerWithWebhookDocument {
   timers: Record<string, CrdtTimer>;
   webhookConfigs: Record<string, WebhookConfigShape>;
+}
+
+/**
+ * Helper to build a document carrying both timers and webhookConfigs — used
+ * by the cross-validator runAllValidations test that exercises the
+ * "timer error does not block webhook" path.
+ */
+function createTimerAndWebhookSession(
+  keys: DocumentKeys,
+  docId: string,
+): EncryptedSyncSession<TimerWithWebhookDocument> {
+  const base = fromDoc({ timers: {}, webhookConfigs: {} });
+  return new EncryptedSyncSession<TimerWithWebhookDocument>({
+    doc: Automerge.clone(base),
+    keys,
+    documentId: asSyncDocId(docId),
+    sodium,
+  });
 }
 
 function makeWebhookConfig(url: string, eventTypes: string[], enabled = true): WebhookConfigShape {
@@ -372,227 +367,9 @@ describe("normalizeTimerConfig (M4 extended coverage)", () => {
   });
 });
 
-// ── normalizeWebhookConfigs ──────────────────────────────────────────
+// ── runAllValidations integration: timer paths ───────────────────────
 
-describe("normalizeWebhookConfigs (M4)", () => {
-  let keys: DocumentKeys;
-
-  beforeEach(() => {
-    keys = makeKeys();
-  });
-
-  it("generates notification for config with invalid URL format", () => {
-    const session = createSessionWithWebhookConfigs(keys, "doc-m4-wh-badurl");
-
-    session.change((d) => {
-      d.webhookConfigs["wh_1"] = makeWebhookConfig("not-a-url", ["member.created"]);
-    });
-
-    const result = normalizeWebhookConfigs(session);
-
-    expect(result.count).toBe(1);
-    expect(result.notifications).toHaveLength(1);
-    expect(result.notifications[0]?.fieldName).toBe("url");
-    expect(result.notifications[0]?.resolution).toBe("notification-only");
-    expect(result.notifications[0]?.summary).toContain("invalid URL format");
-    expect(result.envelope).toBeNull();
-  });
-
-  it("generates notification for config with non-HTTP(S) URL (ftp://)", () => {
-    const session = createSessionWithWebhookConfigs(keys, "doc-m4-wh-ftp");
-
-    session.change((d) => {
-      d.webhookConfigs["wh_1"] = makeWebhookConfig("ftp://example.com/hook", ["member.created"]);
-    });
-
-    const result = normalizeWebhookConfigs(session);
-
-    expect(result.count).toBe(1);
-    expect(result.notifications).toHaveLength(1);
-    expect(result.notifications[0]?.fieldName).toBe("url");
-    expect(result.notifications[0]?.summary).toContain("non-HTTP(S) URL");
-  });
-
-  it("produces no notification for config with valid HTTPS URL", () => {
-    const session = createSessionWithWebhookConfigs(keys, "doc-m4-wh-https");
-
-    session.change((d) => {
-      d.webhookConfigs["wh_1"] = makeWebhookConfig("https://example.com/webhook", [
-        "member.created",
-      ]);
-    });
-
-    const result = normalizeWebhookConfigs(session);
-
-    expect(result.count).toBe(0);
-    expect(result.notifications).toHaveLength(0);
-  });
-
-  it("produces no notification for config with valid HTTP URL", () => {
-    const session = createSessionWithWebhookConfigs(keys, "doc-m4-wh-http");
-
-    session.change((d) => {
-      d.webhookConfigs["wh_1"] = makeWebhookConfig("http://localhost:8080/hook", [
-        "member.created",
-      ]);
-    });
-
-    const result = normalizeWebhookConfigs(session);
-
-    expect(result.count).toBe(0);
-    expect(result.notifications).toHaveLength(0);
-  });
-
-  it("skips URL validation when url is a plain string (non-object guard)", () => {
-    const session = createSessionWithWebhookConfigs(keys, "doc-m4-wh-plain-url");
-
-    session.change((d) => {
-      d.webhookConfigs["wh_1"] = makeWebhookConfig("https://example.com/hook", ["member.created"]);
-    });
-
-    const result = normalizeWebhookConfigs(session);
-
-    // URL validation runs normally — valid HTTPS URL produces no notification
-    expect(result.count).toBe(0);
-    expect(result.notifications).toHaveLength(0);
-  });
-
-  it("generates notification for config with unknown event type", () => {
-    const session = createSessionWithWebhookConfigs(keys, "doc-m4-wh-unk-evt");
-
-    session.change((d) => {
-      d.webhookConfigs["wh_1"] = makeWebhookConfig("https://example.com/hook", [
-        "totally.fake.event",
-      ]);
-    });
-
-    const result = normalizeWebhookConfigs(session);
-
-    expect(result.count).toBe(1);
-    expect(result.notifications).toHaveLength(1);
-    expect(result.notifications[0]?.fieldName).toBe("eventTypes");
-    expect(result.notifications[0]?.summary).toContain("unknown event type");
-  });
-
-  it("produces no notification for config with all valid event types", () => {
-    const session = createSessionWithWebhookConfigs(keys, "doc-m4-wh-valid-evt");
-
-    session.change((d) => {
-      d.webhookConfigs["wh_1"] = makeWebhookConfig("https://example.com/hook", [
-        "member.created",
-        "fronting.started",
-        "fronting.ended",
-      ]);
-    });
-
-    const result = normalizeWebhookConfigs(session);
-
-    expect(result.count).toBe(0);
-    expect(result.notifications).toHaveLength(0);
-  });
-
-  it("validates plain-string event types via typeof fallback", () => {
-    const session = createSessionWithWebhookConfigs(keys, "doc-m4-wh-plain-str-evt");
-
-    session.change((d) => {
-      d.webhookConfigs["wh_1"] = makeWebhookConfig("https://example.com/hook", [
-        "member.created",
-        "fronting.started",
-      ]);
-    });
-
-    const result = normalizeWebhookConfigs(session);
-
-    // Valid event types handled via ImmutableString .val extraction
-    expect(result.count).toBe(0);
-    expect(result.notifications).toHaveLength(0);
-  });
-
-  it("generates one notification per config for mix of valid and invalid event types (breaks on first invalid)", () => {
-    const session = createSessionWithWebhookConfigs(keys, "doc-m4-wh-mix-evt");
-
-    session.change((d) => {
-      d.webhookConfigs["wh_1"] = makeWebhookConfig("https://example.com/hook", [
-        "member.created",
-        "fake.event.one",
-        "fake.event.two",
-      ]);
-    });
-
-    const result = normalizeWebhookConfigs(session);
-
-    expect(result.count).toBe(1);
-    expect(result.notifications).toHaveLength(1);
-    expect(result.notifications[0]?.fieldName).toBe("eventTypes");
-  });
-
-  it("returns count=0 for empty configs map", () => {
-    const session = createSessionWithWebhookConfigs(keys, "doc-m4-wh-empty");
-
-    const result = normalizeWebhookConfigs(session);
-
-    expect(result.count).toBe(0);
-    expect(result.notifications).toHaveLength(0);
-    expect(result.envelope).toBeNull();
-  });
-
-  it("returns count=0 for config with empty eventTypes array", () => {
-    const session = createSessionWithWebhookConfigs(keys, "doc-m4-wh-empty-evt");
-
-    session.change((d) => {
-      d.webhookConfigs["wh_1"] = makeWebhookConfig("https://example.com/hook", []);
-    });
-
-    const result = normalizeWebhookConfigs(session);
-
-    expect(result.count).toBe(0);
-    expect(result.notifications).toHaveLength(0);
-  });
-
-  it("skips eventTypes validation when eventTypes is not an array", () => {
-    const session = createSessionWithWebhookConfigs(keys, "doc-m4-wh-nonarr-evt");
-
-    session.change((d) => {
-      d.webhookConfigs["wh_1"] = {
-        url: s("https://example.com/hook"),
-        // @ts-expect-error -- deliberately setting eventTypes to a non-array to test guard
-        eventTypes: "not-an-array",
-        enabled: true,
-      };
-    });
-
-    const result = normalizeWebhookConfigs(session);
-
-    // Array.isArray guard skips the eventTypes loop
-    expect(result.count).toBe(0);
-    expect(result.notifications).toHaveLength(0);
-  });
-
-  it("only invalid configs generate notifications in a multi-config set", () => {
-    const session = createSessionWithWebhookConfigs(keys, "doc-m4-wh-multi");
-
-    session.change((d) => {
-      d.webhookConfigs["wh_valid"] = makeWebhookConfig("https://example.com/hook", [
-        "member.created",
-      ]);
-      d.webhookConfigs["wh_bad_url"] = makeWebhookConfig("not-a-valid-url", ["member.created"]);
-      d.webhookConfigs["wh_bad_evt"] = makeWebhookConfig("https://example.com/other", [
-        "nonexistent.event",
-      ]);
-    });
-
-    const result = normalizeWebhookConfigs(session);
-
-    expect(result.count).toBe(2);
-    expect(result.notifications).toHaveLength(2);
-    const validConfigNotifications = result.notifications.filter((n) => n.entityId === "wh_valid");
-    expect(validConfigNotifications).toHaveLength(0);
-  });
-});
-
-// ── runAllValidations integration (M4 fields) ───────────────────────
-
-describe("runAllValidations (M4 timer + webhook integration)", () => {
+describe("runAllValidations (M4 timer integration)", () => {
   let keys: DocumentKeys;
 
   beforeEach(() => {
@@ -615,40 +392,8 @@ describe("runAllValidations (M4 timer + webhook integration)", () => {
     expect(session.document.timers[asTimerId("tmr_1")]?.enabled).toBe(false);
   });
 
-  it("triggers normalizeWebhookConfigs when document has webhookConfigs field", () => {
-    const session = createSessionWithWebhookConfigs(keys, "doc-m4-run-webhook");
-
-    session.change((d) => {
-      d.webhookConfigs["wh_1"] = makeWebhookConfig("ftp://bad-protocol.com/hook", [
-        "member.created",
-      ]);
-    });
-
-    const result = runAllValidations(session);
-
-    expect(result.webhookConfigIssues).toBe(1);
-    expect(result.notifications.some((n) => n.entityType === "webhook-config")).toBe(true);
-  });
-
-  it("runs both validators when both timers and webhookConfigs are present", () => {
-    const session = createSessionWithWebhookConfigs(keys, "doc-m4-run-both");
-
-    session.change((d) => {
-      d.timers[asTimerId("tmr_1")] = makeTimer("tmr_1", { intervalMinutes: -5 });
-      d.webhookConfigs["wh_1"] = makeWebhookConfig("not-valid", ["member.created"]);
-    });
-
-    const result = runAllValidations(session);
-
-    expect(result.timerConfigNormalizations).toBe(1);
-    expect(result.webhookConfigIssues).toBe(1);
-    expect(result.notifications.some((n) => n.resolution === "post-merge-timer-normalize")).toBe(
-      true,
-    );
-    expect(result.notifications.some((n) => n.entityType === "webhook-config")).toBe(true);
-  });
-
   it("runs neither validator when neither field is present in the document", () => {
+    // Cross-coverage: also implicitly tests the no-webhookConfigs branch.
     const base = createFrontingDocument();
     const session = new EncryptedSyncSession({
       doc: Automerge.clone(base),
@@ -664,7 +409,9 @@ describe("runAllValidations (M4 timer + webhook integration)", () => {
   });
 
   it("catches timer validator error without blocking webhook validation", () => {
-    const session = createSessionWithWebhookConfigs(keys, "doc-m4-run-error");
+    // Cross-coverage: this exercises both validators — the timer-error path
+    // is the test focus; the webhook validator runs as the success companion.
+    const session = createTimerAndWebhookSession(keys, "doc-m4-run-error");
 
     session.change((d) => {
       // @ts-expect-error -- deliberately corrupting the timers map to test error handling
@@ -687,34 +434,5 @@ describe("runAllValidations (M4 timer + webhook integration)", () => {
     // Webhook validator should still have run successfully
     expect(result.webhookConfigIssues).toBe(1);
     expect(result.notifications.some((n) => n.entityType === "webhook-config")).toBe(true);
-  });
-
-  it("catches webhook validator error without blocking timer validation", () => {
-    const session = createSessionWithWebhookConfigs(keys, "doc-m4-run-error-rev");
-
-    session.change((d) => {
-      d.timers[asTimerId("tmr_1")] = makeTimer("tmr_1", { intervalMinutes: -5 });
-      // @ts-expect-error -- deliberately corrupting the webhookConfigs map to test error handling
-      d.webhookConfigs["wh_broken"] = null;
-    });
-
-    const errorMessages: string[] = [];
-    const result = runAllValidations(session, (msg) => {
-      errorMessages.push(msg);
-    });
-
-    // Guard: ensure the error path actually triggered
-    expect(result.errors.length).toBeGreaterThan(0);
-
-    const webhookError = result.errors.find((e) => e.validator === "normalizeWebhookConfigs");
-    expect(webhookError?.validator).toBe("normalizeWebhookConfigs");
-    expect(errorMessages).toContain("Webhook config validation failed");
-
-    // Timer validator should still have run successfully
-    expect(result.timerConfigNormalizations).toBe(1);
-    expect(result.notifications.some((n) => n.resolution === "post-merge-timer-normalize")).toBe(
-      true,
-    );
-    expect(session.document.timers[asTimerId("tmr_1")]?.enabled).toBe(false);
   });
 });
